requirements.txt
python-dotenv~=0.21.0
requests~=2.27.1
lxml >= 4.3.0
pyVmomi >= 6.7
six (>=1.12)

requests (==2.27.1)
pyOpenSSL (==22.0.0)
cryptography (==36.0.0)
setuptools (==62.0.0)
###### SDK requirements ######
vapi-client-bindings == 4.0.0
vmc-client-bindings
nsx-python-sdk
nsx-policy-python-sdk
nsx-vmc-policy-python-sdk
nsx-vmc-aws-integration-python-sdk
vmc-draas-client-bindings